requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-jazzmin==2.6.1
django-js-asset==2.2.0
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1

django-modeltranslation~=0.18.11
typing_extensions~=4.10.0
pip~=24.0
numpy~=1.26.4